fixes and dosc

+++ tomato.sh
@@ -1,31 +1,51 @@
 #!/bin/bash
 
+# Copyright (C) 2015 Szymon Kopciewski
+#
+# This file is part of Tomato
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
 IDLE_PID_FILE='/tmp/tomato-idle.pid'
 ACTIVE_PID_FILE='/tmp/tomato-active.pid'
 TIMER_PID_FILE='/tmp/tomato-timer.pid'
 SCRIPT_PATH=`realpath $0`
+TRAYTO_PATH=`dirname $SCRIPT_PATH`/vendors/traytor/bin/traytor
 
 CONFIG=~/.tomato/config
+### Configurable through config file ###
 TIMER_MESSAGE=' !! Take a BREAK !! '
 TIMER_MINUTES=25
 TIMER_CALLBACK='true'
 IDLE_ICON=~/.tomato/tomato_idle.png
 ACTIVE_ICON=~/.tomato/tomato_active.png
+########################################
 
 function run_tomato_idle
 {
-  (/home/szef/sbin/traytor \
+  ($TRAYTO_PATH \
    -t "Tomato: idle" \
-   -c "$SCRIPT_PATH" \
+   -c "$SCRIPT_PATH &> /dev/null" \
    $IDLE_ICON) &
   save_pid_to_file $! $IDLE_PID_FILE
 }
 
 function run_tomato_avtive
 {
-  (/home/szef/sbin/traytor \
+  ($TRAYTO_PATH \
     -t "Tomato: active" \
-    -c "$SCRIPT_PATH" \
+    -c "$SCRIPT_PATH &> /dev/null" \
     $ACTIVE_ICON) &
   save_pid_to_file $! $ACTIVE_PID_FILE
 }
@@ -35,7 +55,7 @@ function run_tomato_timer
   local seconds=$((60 * $TIMER_MINUTES))
   (sleep $seconds && \
     notify-send "$TIMER_MESSAGE" && \
-    $($TIMER_CALLBACK) && \
+    ($TIMER_CALLBACK) && \
     ($SCRIPT_PATH)& ) &
   save_pid_to_file $! $TIMER_PID_FILE
 }
@@ -55,6 +75,17 @@ function stop_tomato_timer
   stop_process_from_file $TIMER_PID_FILE
 }
 
+function is_tomato_idle
+{
+  check_process_from_file $IDLE_PID_FILE
+}
+
+function is_tomato_active
+{
+  check_process_from_file $TIMER_PID_FILE || \
+    check_process_from_file $ACTIVE_PID_FILE
+}
+
 function save_pid_to_file
 {
   local pid=$1
@@ -65,6 +96,8 @@ function save_pid_to_file
 function stop_process_from_file
 {
   local file=$1
+  [ -f $file ] || return
+  pkill -P `cat $file` &> /dev/null
   pkill -F $file &> /dev/null
   rm -f $file
 }
@@ -79,16 +112,6 @@ function check_process_from_file {
   return 0
 }
 
-function is_tomato_idle
-{
-  check_process_from_file $IDLE_PID_FILE
-}
-
-function is_tomato_active
-{
-  check_process_from_file $ACTIVE_PID_FILE
-}
-
 function prepare_default_config
 {
   if [ ! -f $CONFIG ]; then
@@ -96,7 +119,7 @@ function prepare_default_config
 	cat <<- EOF > $CONFIG
 		# TIMER_MINUTES=$TIMER_MINUTES
 		# TIMER_MESSAGE='$TIMER_MESSAGE'
-		# TIMER_CALLBACK=$TIMER_CALLBACK
+		# TIMER_CALLBACK='$TIMER_CALLBACK'
 		# IDLE_ICON=$IDLE_ICON
 		# ACTIVE_ICON=$ACTIVE_ICON
 	EOF
